chore(init): move the workspace redirect out of run.ts
